Optimize
Make the dictionaries passed to generateCodeWithSaltUnsafe just use key->value instead of (id->key) -> (string->value). So this gives the ability to fill the dictionaries with more than just a single entry.
Optimized the inner loop of 'reveal:' to not call objc functions inside it and removed the additional loop counter.
Some other slight cleanups.
Make the example project point to the parent directories Obfuscator folder.
Updated docs for the dictionary change.
Let Xcode 12 update the project with all the usual noise and changed the deployment target to iOS 9.