Add a CLI
Had to modify the Obfuscator.h and m files to return their dictionaries/results. Removed the DEBUG macros so this thing works in release mode.
Optimized the runtime 'reveal:' again to remove another loop counter and obj-c call, it's as tight as can be now. Not maximum tightness but a good amount of tightness.
Changed all (unsigned char *) to (char *) throughout project, including the code it produces.
Removed 'CStringToNSString' and 'convertNSStringtoCString', we can use @(boxing) and UTF8String methods.
@synchronized the saltDatabase for thread safety.