Optimize build times for large salts
After a winning permutation is discovered, we may as well reuse it for subsequent builds.
This update allows passing a modified version of the salt back where it is checked first in the permutation list.
enumerateAllPermutationsOfArray is a new enumeration method, returning each permutation one at a time through a block callback, so every permutation doesn't have to be precalculated into memory which was taking inordinate amounts of time and space.
generateCodeWithSalt modified so that it can now succeed on the first permutation if it's a winner.